+++ my-rpc-core/src/main/java/github/qiao712/rpc/transport/RpcServer.java
@@ -14,7 +14,7 @@ public interface RpcServer {
 
     SerializationType getSerializationType();
 
-    int getMaxIdleTime();
+    long getMaxIdleTime();
 
-    void setMaxIdleTime(int maxIdleTime);
+    void setMaxIdleTime(long maxIdleTime);
 }
